lesson 5 hash tokens authorization

+++ src/services/user.service.ts
@@ -14,18 +14,29 @@ class UserService {
     public create(user: IUserCreateDTO): Promise<IUser> {
         return userRepository.create(user);
     }
-    public getById(userId: string): Promise<IUser> {
-        const user = userRepository.getById(userId);
+    public async getById(userId: string): Promise<IUser> {
+        const user = await userRepository.getById(userId);
         if (!user) {
             throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
         }
         return user;
     }
-    public updateById(userId: string, user: IUserUpdateDTO): Promise<IUser> {
-        return userRepository.updateById(userId, user);
+    public async updateById(
+        userId: string,
+        user: IUserUpdateDTO,
+    ): Promise<IUser> {
+        const data = await userRepository.getById(userId);
+        if (!data) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return await userRepository.updateById(userId, user);
     }
-    public deleteById(userId: string): Promise<void> {
-        return userRepository.deleteById(userId);
+    public async deleteById(userId: string): Promise<void> {
+        const data = await userRepository.getById(userId);
+        if (!data) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return await userRepository.deleteById(userId);
     }
     public async isEmailUnique(email: string): Promise<void> {
         const user = await userRepository.getByEmail(email);
